chore: 文件操作

+++ index.js
@@ -6,11 +6,28 @@ import camelcase from "camelcase";
 export default async () => {
   const root = path.join(process.cwd(), "events");
   mkdirp.sync(root);
-  const index = path.join(root, "index.js");
+
   const ignore = path.join(root, ".gitignore");
-  fs.writeFileSync(ignore, "/index.js");
+  fs.stat(ignore, function (err, stat) {
+    if (stat && stat.isFile()) {
+      console.log(`${ignore}存在`);
+    } else {
+      fs.writeFileSync(ignore, "/index.js");
+    }
+  });
+
   const lib = path.join(root, "lib");
   mkdirp.sync(lib);
+
+  const application = path.join(lib, "application.js");
+  fs.stat(application, function (err, stat) {
+    if (stat && stat.isFile()) {
+      console.log(`${application}存在`);
+    } else {
+      fs.writeFileSync(application, "export default ['didFinishLaunching']");
+    }
+  });
+
   const files = fs.readdirSync(lib);
   const out = {};
   for (const file of files) {
@@ -28,6 +45,8 @@ export default async () => {
   let json = JSON.stringify(out);
   json = json.replaceAll('$"', "");
   json = json.replaceAll('"$', "");
+
+  const index = path.join(root, "index.js");
   fs.writeFileSync(
     index,
     'import EventEmitter from "events"; export const emitter = new EventEmitter(); export default ' +
